+++ tests/testthat/test-broom.R
@@ -205,54 +205,54 @@ nmTest({
         td <- broom.mixed::tidy(fitF, effects = ef, exponentiate = NA)
         td1 <- td$estimate
         check_tidy(td, 36, 5, c("effect", "group", "level", "term", "estimate"))
-        ##
+
         td <- broom.mixed::tidy(fitF, effects = ef, exponentiate = FALSE)
         td2 <- td$estimate
         check_tidy(td, 36, 5, c("effect", "group", "level", "term", "estimate"))
-        ##
+
         td <- broom.mixed::tidy(fitF, effects = ef, exponentiate = TRUE)
         td3 <- td$estimate
         check_tidy(td, 36, 5, c("effect", "group", "level", "term", "estimate"))
-        ##
+
         expect_equal(td1, td2)
         expect_equal(td2, td3)
       }
 
-      ##
+
       td <- broom.mixed::tidy(fitF, effects = "ran_coef", exponentiate = NA)
       td1 <- td$estimate
       check_tidy(td, 36, 5, c("effect", "group", "level", "term", "estimate"))
-      ##
+
       td <- broom.mixed::tidy(fitF, effects = "ran_coef", exponentiate = FALSE)
       td2 <- td$estimate
       check_tidy(td, 36, 5, c("effect", "group", "level", "term", "estimate"))
-      ##
+
       td <- broom.mixed::tidy(fitF, effects = "ran_coef", exponentiate = TRUE)
       td3 <- td$estimate
       check_tidy(td, 36, 5, c("effect", "group", "level", "term", "estimate"))
-      ##
+
       expect_equal(log(td1), td2)
       expect_equal(td2, log(td3))
-      ##
+
       td <- broom.mixed::tidy(fitF, effects = "ran_pars", exponentiate = NA)
       td1 <- td$estimate
       check_tidy(td, 4, 4, c("effect", "group", "term", "estimate"))
-      ##
+
       td <- broom.mixed::tidy(fitF, effects = "ran_pars", exponentiate = FALSE)
       td2 <- td$estimate
       check_tidy(td, 4, 4, c("effect", "group", "term", "estimate"))
-      ##
+
       td <- broom.mixed::tidy(fitF, effects = "ran_pars", exponentiate = TRUE)
       td3 <- td$estimate
       check_tidy(td, 4, 4, c("effect", "group", "term", "estimate"))
-      ##
+
       expect_equal(td1, td2, tolerance = tol)
       expect_equal(td2, td3, tolerance = tol)
     })
   }
 
   for (f in c("foi", "fo")) {
-    fitF <- suppressMessages(suppressWarnings(nlmixr(one.compartment, theo_sd, est = f, control=list(print=0))))
+    fitF <- .nlmixr(one.compartment, theo_sd, est = f, control=list(print=0))
     test_that(sprintf("tidy works on nlmixr fit %s fits", f), {
       td <- broom.mixed::tidy(fitF, exponentiate = NA)
       check_tidy(td, 7, 7, c("effect", "group", "term", "estimate", "std.error", "statistic", "p.value"))
@@ -299,52 +299,52 @@ nmTest({
         td <- broom.mixed::tidy(fitF, effects = ef, exponentiate = NA)
         td1 <- td$estimate
         check_tidy(td, 36, 5, c("effect", "group", "level", "term", "estimate"))
-        ##
+
         td <- broom.mixed::tidy(fitF, effects = ef, exponentiate = FALSE)
         td2 <- td$estimate
         check_tidy(td, 36, 5, c("effect", "group", "level", "term", "estimate"))
-        ##
+
         td <- broom.mixed::tidy(fitF, effects = ef, exponentiate = TRUE)
         td3 <- td$estimate
         check_tidy(td, 36, 5, c("effect", "group", "level", "term", "estimate"))
-        ##
+
         expect_equal(td1, td2, tolerance = tol)
         expect_equal(td2, td3, tolerance = tol)
       }
-      ##
+
       td <- broom.mixed::tidy(fitF, effects = "ran_coef", exponentiate = NA)
       td1 <- td$estimate
       check_tidy(td, 36, 5, c("effect", "group", "level", "term", "estimate"))
-      ##
+
       td <- broom.mixed::tidy(fitF, effects = "ran_coef", exponentiate = FALSE)
       td2 <- td$estimate
       check_tidy(td, 36, 5, c("effect", "group", "level", "term", "estimate"))
-      ##
+
       td <- broom.mixed::tidy(fitF, effects = "ran_coef", exponentiate = TRUE)
       td3 <- td$estimate
       check_tidy(td, 36, 5, c("effect", "group", "level", "term", "estimate"))
-      ##
+
       expect_equal(log(td1), td2, tolerance = tol)
       expect_equal(td2, log(td3), tolerance = tol)
-      ##
+
       td <- broom.mixed::tidy(fitF, effects = "ran_pars", exponentiate = NA)
       td1 <- td$estimate
       check_tidy(td, 4, 4, c("effect", "group", "term", "estimate"))
-      ##
+
       td <- broom.mixed::tidy(fitF, effects = "ran_pars", exponentiate = FALSE)
       td2 <- td$estimate
       check_tidy(td, 4, 4, c("effect", "group", "term", "estimate"))
-      ##
+
       td <- broom.mixed::tidy(fitF, effects = "ran_pars", exponentiate = TRUE)
       td3 <- td$estimate
       check_tidy(td, 4, 4, c("effect", "group", "term", "estimate"))
-      ##
+
       expect_equal(td1, td2, tolerance = tol)
       expect_equal(td2, td3, tolerance = tol)
     })
   }
 
-  fitP <- suppressMessages(suppressWarnings(nlmixr(one.compartment, theo_sd, est = "posthoc")))
+  fitP <- .nlmixr(one.compartment, theo_sd, est = "posthoc")
 
   test_that("tidy works on posthoc fit fits", {
 
@@ -457,5 +457,4 @@ nmTest({
     expect_equal(td1, td2, tolerance = tol)
     expect_equal(td2, td3, tolerance = tol)
   })
-
 })