+++ tests/testthat/test-manual-back-transform.R
@@ -35,13 +35,11 @@ nmTest({
     expect_equal(fit$parFixed["tka", "Back-transformed(95%CI)"],
                  sprintf("%3g", fit$parFixedDf["tka", "Estimate"] * 100))
 
-    expect_equal(NA_real_,
-                 setNames(fit$parFixedDf["tka", "CI Lower"], NULL))
-    expect_equal(NA_real_,
-                 setNames(fit$parFixedDf["tka", "CI Upper"], NULL))
+  expect_equal(fit$parFixed["tka", "Back-transformed(95%CI)"],
+               formatMinWidth(fit$parFixedDf["tka", "Estimate"] * 100))
 
-    expect_equal(setNames(fit$parFixedDf["tcl", "Estimate"] * 100, NULL),
-                 setNames(fit$parFixedDf["tcl", "Back-transformed"], NULL))
+  expect_equal(setNames(fit$parFixedDf["tcl", "Estimate"] * 100, NULL),
+               setNames(fit$parFixedDf["tcl", "Back-transformed"], NULL))
 
 
     qn <- qnorm(1.0 - (1 - 0.95) / 2)
@@ -52,11 +50,11 @@ nmTest({
     expect_equal(setNames(t100(fit$parFixedDf["tcl", "Estimate"] - qn * fit$parFixedDf["tcl", "SE"]), NULL),
                  setNames(fit$parFixedDf["tcl", "CI Lower"], NULL))
 
-    expect_equal(fit$parFixed["tcl", "Back-transformed(95%CI)"],
-                 sprintf("%3g (%3g, %3g)",
-                         t100(fit$parFixedDf["tcl", "Estimate"]),
-                         setNames(t100(fit$parFixedDf["tcl", "Estimate"] - qn * fit$parFixedDf["tcl", "SE"]), NULL),
-                         setNames(t100(fit$parFixedDf["tcl", "Estimate"] + qn * fit$parFixedDf["tcl", "SE"]), NULL)))
+  expect_equal(fit$parFixed["tcl", "Back-transformed(95%CI)"],
+               sprintf("%s (%s, %s)",
+                       formatMinWidth(t100(fit$parFixedDf["tcl", "Estimate"])),
+                       formatMinWidth(t100(fit$parFixedDf["tcl", "Estimate"] - qn * fit$parFixedDf["tcl", "SE"])),
+                       formatMinWidth(t100(fit$parFixedDf["tcl", "Estimate"] + qn * fit$parFixedDf["tcl", "SE"]))))
 
 
     expect_equal(setNames(exp(fit$parFixedDf["tv", "Estimate"] + qn * fit$parFixedDf["tv", "SE"]), NULL),
@@ -76,8 +74,8 @@ nmTest({
     expect_equal(setNames(fit$parFixedDf["tka", "Estimate"] * 100, NULL),
                  setNames(fit$parFixedDf["tka", "Back-transformed"], NULL))
 
-    expect_equal(fit$parFixed["tka", "Back-transformed(80%CI)"],
-                 sprintf("%3g", fit$parFixedDf["tka", "Estimate"] * 100))
+  expect_equal(fit$parFixed["tka", "Back-transformed(80%CI)"],
+               formatMinWidth(fit$parFixedDf["tka", "Estimate"] * 100))
 
     expect_equal(NA_real_,
                  setNames(fit$parFixedDf["tka", "CI Lower"], NULL))
@@ -94,11 +92,11 @@ nmTest({
     expect_equal(setNames(t100(fit$parFixedDf["tcl", "Estimate"] - qn * fit$parFixedDf["tcl", "SE"]), NULL),
                  setNames(fit$parFixedDf["tcl", "CI Lower"], NULL))
 
-    expect_equal(fit$parFixed["tcl", "Back-transformed(80%CI)"],
-                 sprintf("%3g (%3g, %3g)",
-                         t100(fit$parFixedDf["tcl", "Estimate"]),
-                         setNames(t100(fit$parFixedDf["tcl", "Estimate"] - qn * fit$parFixedDf["tcl", "SE"]), NULL),
-                         setNames(t100(fit$parFixedDf["tcl", "Estimate"] + qn * fit$parFixedDf["tcl", "SE"]), NULL)))
+  expect_equal(fit$parFixed["tcl", "Back-transformed(80%CI)"],
+               sprintf("%s (%s, %s)",
+                       formatMinWidth(t100(fit$parFixedDf["tcl", "Estimate"])),
+                       formatMinWidth(t100(fit$parFixedDf["tcl", "Estimate"] - qn * fit$parFixedDf["tcl", "SE"])),
+                       formatMinWidth(t100(fit$parFixedDf["tcl", "Estimate"] + qn * fit$parFixedDf["tcl", "SE"]))))
 
 
     expect_equal(setNames(exp(fit$parFixedDf["tv", "Estimate"] + qn * fit$parFixedDf["tv", "SE"]), NULL),